Using bb8 pool in lipl-postgres-repo

+++ crates/lipl-postgres-repo/src/constant.rs
@@ -1,1 +1,1 @@
-pub const POOL_MAX_SIZE: usize = 16;
+pub const POOL_MAX_SIZE: u32 = 16;
